Fixed status codes

+++ app/controllers/controllers.js
@@ -26,8 +26,8 @@ module.exports.log_in = (req, res) => {
             token: token
         })
     }, () => {
-        res.status(400).send({
-            massage: "username or password is not correct!"
+        res.status(401).send({
+            massage: "Username or password is not correct!"
         })
     });
 
@@ -50,7 +50,7 @@ module.exports.validate_token = (req, res) => {
             username: user.username
         })
     }, () => {
-        res.status(400).send({
+        res.status(401).send({
             massage: "Token is incorrect or expired"
         })
     });
@@ -75,53 +75,58 @@ module.exports.get_map = (req, res) => {
 
 //post buses location.
 module.exports.post_location = (req, res) => {
+    database.checkToken(req.header("token"), (result) => {
+        let test = true;
+        let imei = req.params.imei;
+        let q = url.parse(req.url, true).query;
 
-    let test = true;
-    let imei = req.params.imei;
-    let q = url.parse(req.url, true).query;
-
-    // Validate request
-    if (!req.body) {
-        test = false;
-        res.status(400).send({
-            message: "Content can not be empty!"
-        });
-    }
-
-    if (!check.posted_location(imei, q, database.buses)) {
-        test = false;
-        res.status(400).send({
-            message: "Content structure is not correct!"
-        });
-    }
-
-    if (test) {
-        let i = 0;
-        let bus;
-        for (let i = 0; i < database.buses.length; i++) {
-            if (imei == database.buses[i].imei) {
-                bus = database.buses[i];
-                bus.loc.long = q.longitude;
-                bus.loc.lat = q.latitude;
-                bus.time = Math.round(new Date().getTime() / 1000);
-                database.updateBusInfo(bus);
-                res.status(200).send({
-                    message: "DONE."
-                });
-                break;
-            }
+        // Validate request
+        if (!req.body) {
+            test = false;
+            res.status(400).send({
+                message: "Content can not be empty!"
+            });
         }
 
-        for (let j = 0; j < database.lines.length; j++) {
-            if (database.lines[j].index == bus.line) {
-                if (!check.in_line(q.latitude, q.longitude, lines[j].map)) {
-                    /*########################################################################################
-                    fire the alert.
-                    ########################################################################################*/
+        if (!check.posted_location(imei, q, database.buses)) {
+            test = false;
+            res.status(400).send({
+                message: "Content structure is not correct!"
+            });
+        }
+
+        if (test) {
+            let i = 0;
+            let bus;
+            for (let i = 0; i < database.buses.length; i++) {
+                if (imei == database.buses[i].imei) {
+                    bus = database.buses[i];
+                    bus.loc.long = q.longitude;
+                    bus.loc.lat = q.latitude;
+                    bus.time = Math.round(new Date().getTime() / 1000);
+                    database.updateBusInfo(bus);
+                    res.status(200).send({
+                        message: "DONE."
+                    });
+                    break;
+                }
+            }
+
+            for (let j = 0; j < database.lines.length; j++) {
+                if (database.lines[j].index == bus.line) {
+                    if (!check.in_line(q.latitude, q.longitude, lines[j].map)) {
+                        /*########################################################################################
+                        fire the alert.
+                        ########################################################################################*/
+                    }
                 }
             }
         }
-    }
+    }, () => {
+        res.status(401).send({
+            message: "Access Denied"
+        });
+    });
 }
 
 //..................................................................
@@ -129,55 +134,62 @@ module.exports.post_location = (req, res) => {
 
 //Add a new line.
 module.exports.add_line = (req, res) => {
-    let q = req.params;
-    let test = true;
-    let line_c = {
-        name: '',
-        map: [],
-        index: null,
-        stops: []
-    }
-    // Validate request
-    if (!req.body) {
-        test = false;
-        res.status(400).send({
-            message: "Content can not be empty!"
-        });
-    }
-
-    if (!check.line_check(q.name, req.body.map, req.body.stops)) {
-        test = false;
-        res.status(400).send({
-            message: "Content structure is not correct!"
-        });
-    }
-
-    if (!check.line_is_new(q.name, database.lines)) {
-        test = false;
-        res.status(400).send({
-            message: "Line olready exist!"
-        });
-    }
     database.checkToken(req.header("token"), (result) => {
-        if (test) {
-            line_c.name = q.name;
-            line_c.map = req.body.map;
-            line_c.stops = req.body.stops;
-            if (database.lines.length == 0) {
-                indx = 0;
+        if (result.role === 'admin') {
+            let q = req.params;
+            let test = true;
+            let line_c = {
+                name: '',
+                map: [],
+                index: null,
+                stops: []
             }
-            else {
-                indx = database.lines[database.lines.length - 1].index;
+            // Validate request
+            if (!req.body) {
+                test = false;
+                res.status(400).send({
+                    message: "Content can not be empty!"
+                });
             }
-            indx++;
-            line_c.index = indx;
-            database.addLine(line_c);
-            res.status(200).send({
-                message: "index:" + line_c.index
+
+            if (!check.line_check(q.name, req.body.map, req.body.stops)) {
+                test = false;
+                res.status(400).send({
+                    message: "Content structure is not correct!"
+                });
+            }
+
+            if (!check.line_is_new(q.name, database.lines)) {
+                test = false;
+                res.status(403).send({
+                    message: "Line already exist!"
+                });
+            }
+            if (test) {
+                line_c.name = q.name;
+                line_c.map = req.body.map;
+                line_c.stops = req.body.stops;
+                if (database.lines.length == 0) {
+                    indx = 0;
+                }
+                else {
+                    indx = database.lines[database.lines.length - 1].index;
+                }
+                indx++;
+                line_c.index = indx;
+                database.addLine(line_c);
+                res.status(200).send({
+                    message: "index:" + line_c.index
+                });
+            }
+        }
+        else {
+            res.status(401).send({
+                message: "Access Denied"
             });
         }
     }, () => {
-        res.status(403).send({
+        res.status(401).send({
             message: "Access Denied"
         });
     });
@@ -187,54 +199,61 @@ module.exports.add_line = (req, res) => {
 
 // Add a new bus.
 module.exports.add_bus = (req, res) => {
-    let imei = req.params.imei;
-    let q = url.parse(req.url, true).query;
-    let test = true;
-    let bus_c = {
-        imei: '',
-        driver: '',
-        active: true,
-        loc: {
-            lat: null,
-            long: null
-        },
-        line: '',
-        time: null
-    }
-
-    // Validate request
-    if (!req.body) {
-        test = false;
-        res.status(400).send({
-            message: "Content can not be empty!"
-        });
-    }
-
-    if (!check.bus_check(imei, q.line, database.lines)) {
-        test = false;
-        res.status(400).send({
-            message: "Content structure is not correct!"
-        });
-    }
-
-    if (!check.bus_is_new(imei, database.buses)) {
-        test = false;
-        res.status(400).send({
-            message: "Bus olready exist!"
-        });
-    }
-
     database.checkToken(req.header("token"), (result) => {
-        if (test) {
-            bus_c.imei = imei;
-            bus_c.line = q.line;
-            database.addBus(bus_c);
-            res.status(200).send({
-                message: "DONE."
+        if (result.role === 'admin') {
+            let imei = req.params.imei;
+            let q = url.parse(req.url, true).query;
+            let test = true;
+            let bus_c = {
+                imei: '',
+                driver: '',
+                active: true,
+                loc: {
+                    lat: null,
+                    long: null
+                },
+                line: '',
+                time: null
+            }
+
+            // Validate request
+            if (!req.body) {
+                test = false;
+                res.status(400).send({
+                    message: "Content can not be empty!"
+                });
+            }
+
+            if (!check.bus_check(imei, q.line, database.lines)) {
+                test = false;
+                res.status(400).send({
+                    message: "Content structure is not correct!"
+                });
+            }
+
+            if (!check.bus_is_new(imei, database.buses)) {
+                test = false;
+                res.status(403).send({
+                    message: "Bus already exist!"
+                });
+            }
+
+            if (test) {
+                bus_c.imei = imei;
+                bus_c.line = q.line;
+                database.addBus(bus_c);
+                res.status(200).send({
+                    message: "DONE."
+                });
+            }
+        }
+        else {
+            res.status(401).send({
+                message: "Access Denied"
             });
         }
     }, () => {
-        res.status(403).send({
+        res.status(401).send({
             message: "Access Denied"
         });
     });
@@ -248,45 +267,52 @@ module.exports.add_bus = (req, res) => {
 
 // Remove a line.
 module.exports.remove_line = (req, res) => {
-    let q = req.params;
-
-    let test = true;
-    // Validate request
-    if (!req.body) {
-        test = false;
-        res.status(400).send({
-            message: "Content can not be empty!"
-        });
-    }
-
-    if (check.line_is_new(q.name, database.lines)) {
-        test = false;
-        res.status(400).send({
-            message: "Line does not exist!"
-        });
-    }
-
-    if (check.buses_in_line(q.name, database.buses)) {
-        test = false;
-        res.status(401).send({
-            message: "Remove or reassign the buses in the line first!"
-        });
-    }
     database.checkToken(req.header("token"), (result) => {
-        if (test) {
-            for (let i = 0; i < database.lines.length; i++) {
-                if (q.name == database.lines[i].name) {
-                    let line_c = database.lines[i];
-                    database.removeLine(line_c)
-                    res.status(200).send({
-                        message: "DONE."
-                    });
-                    break;
+        if (result.role === 'admin') {
+            let q = req.params;
+
+            let test = true;
+            // Validate request
+            if (!req.body) {
+                test = false;
+                res.status(400).send({
+                    message: "Content can not be empty!"
+                });
+            }
+
+            if (check.line_is_new(q.name, database.lines)) {
+                test = false;
+                res.status(403).send({
+                    message: "Line does not exist!"
+                });
+            }
+
+            if (check.buses_in_line(q.name, database.buses)) {
+                test = false;
+                res.status(401).send({
+                    message: "Remove or reassign the buses in the line first!"
+                });
+            }
+            if (test) {
+                for (let i = 0; i < database.lines.length; i++) {
+                    if (q.name == database.lines[i].name) {
+                        let line_c = database.lines[i];
+                        database.removeLine(line_c)
+                        res.status(200).send({
+                            message: "DONE."
+                        });
+                        break;
+                    }
                 }
             }
         }
+        else {
+            res.status(401).send({
+                message: "Access Denied"
+            });
+        }
     }, () => {
-        res.status(403).send({
+        res.status(401).send({
             message: "Access Denied"
         });
     });
@@ -297,37 +323,44 @@ module.exports.remove_line = (req, res) => {
 
 // Remove a bus.
 module.exports.remove_bus = (req, res) => {
-    let q = req.params;
-    let test = true;
-    // Validate request
-    if (!req.body) {
-        test = false;
-        res.status(400).send({
-            message: "Content can not be empty!"
-        });
-    }
-
-    if (check.bus_is_new(q.imei, database.buses)) {
-        test = false;
-        res.status(400).send({
-            message: "Bus does not exist!"
-        });
-    }
     database.checkToken(req.header("token"), (result) => {
-        if (test) {
-            for (let i = 0; i < database.buses.length; i++) {
-                if (database.buses[i].imei == q.imei) {
-                    let bus_c = database.buses[i];
-                    database.removeBus(bus_c);
-                    res.status(200).send({
-                        message: "DONE."
-                    });
-                    break;
+        if (result.role === 'admin') {
+            let q = req.params;
+            let test = true;
+            // Validate request
+            if (!req.body) {
+                test = false;
+                res.status(400).send({
+                    message: "Content can not be empty!"
+                });
+            }
+
+            if (check.bus_is_new(q.imei, database.buses)) {
+                test = false;
+                res.status(403).send({
+                    message: "Bus does not exist!"
+                });
+            }
+            if (test) {
+                for (let i = 0; i < database.buses.length; i++) {
+                    if (database.buses[i].imei == q.imei) {
+                        let bus_c = database.buses[i];
+                        database.removeBus(bus_c);
+                        res.status(200).send({
+                            message: "DONE."
+                        });
+                        break;
+                    }
                 }
             }
         }
+        else {
+            res.status(401).send({
+                message: "Access Denied"
+            });
+        }
     }, () => {
-        res.status(403).send({
+        res.status(401).send({
             message: "Access Denied"
         });
     });
@@ -338,66 +371,73 @@ module.exports.remove_bus = (req, res) => {
 
 // Assign bus data.
 module.exports.update_bus = (req, res) => {
-    let imei = req.params.imei;
-    let q = url.parse(req.url, true).query;
-    let test = true;
-    let bus_c = {};
-    for (let i = 0; i < database.buses.length; i++) {
-        if (database.buses[i].imei == imei) {
-            bus_c = database.buses[i];
-            break;
-        }
-    }
-
-    // Validate request
-    if (!req.body) {
-        test = false;
-        res.status(400).send({
-            message: "Content can not be empty!"
-        });
-    }
-    else if (check.bus_is_new(imei, database.buses)) {
-        test = false;
-        res.status(400).send({
-            message: "Bus does not exist!"
-        });
-    }
-    else {
-
-        if (q.driver != '') {
-            bus_c.driver = q.driver;
-        }
-        if (q.active != "") {
-            if (q.active == "true") { bus_c.active = true; }
-            else if (q.active == "false") { bus_c.active = false; }
-            else {
-                test = false;
-                res.status(401).send({
-                    message: "active value can only be 'true' or 'false'."
-                });
+    database.checkToken(req.header("token"), (result) => {
+        if (result.role === 'admin') {
+            let imei = req.params.imei;
+            let q = url.parse(req.url, true).query;
+            let test = true;
+            let bus_c = {};
+            for (let i = 0; i < database.buses.length; i++) {
+                if (database.buses[i].imei == imei) {
+                    bus_c = database.buses[i];
+                    break;
+                }
             }
-        }
-        if (q.line != '') {
-            if (check.line_is_new(q.line, database.lines)) {
+
+            // Validate request
+            if (!req.body) {
                 test = false;
                 res.status(400).send({
-                    message: "Line does not exist!"
+                    message: "Content can not be empty!"
+                });
+            }
+            else if (check.bus_is_new(imei, database.buses)) {
+                test = false;
+                res.status(403).send({
+                    message: "Bus does not exist!"
                 });
             }
             else {
-                bus_c.line = q.line;
+
+                if (q.driver != '') {
+                    bus_c.driver = q.driver;
+                }
+                if (q.active != "") {
+                    if (q.active == "true") { bus_c.active = true; }
+                    else if (q.active == "false") { bus_c.active = false; }
+                    else {
+                        test = false;
+                        res.status(401).send({
+                            message: "active value can only be 'true' or 'false'."
+                        });
+                    }
+                }
+                if (q.line != '') {
+                    if (check.line_is_new(q.line, database.lines)) {
+                        test = false;
+                        res.status(400).send({
+                            message: "Line does not exist!"
+                        });
+                    }
+                    else {
+                        bus_c.line = q.line;
+                    }
+                }
+            }
+            if (test) {
+                database.updateBusInfo(bus_c);
+                res.status(200).send({
+                    message: "DONE."
+                });
             }
         }
-    }
-    database.checkToken(req.header("token"), (result) => {
-        if (test) {
-            database.updateBusInfo(bus_c);
-            res.status(200).send({
-                message: "DONE."
+        else {
+            res.status(401).send({
+                message: "Access Denied"
             });
         }
     }, () => {
-        res.status(403).send({
+        res.status(401).send({
             message: "Access Denied"
         });
     });
